aac: rpc param is too long when invoking decoding [1/1]
PD#SWPL-29237

Problem:
RPC param is too long when invoking decoding function

Solution:
Fix it

Verify:
A1

+++ include/aipc_type.h
@@ -99,7 +99,7 @@ typedef struct {
 	uint32_t sampleRate;
 	uint32_t frameSize;
 	uint32_t channels;
-	int channel_counts[0x24];
+	uint8_t channel_counts[0x24];
 } __attribute__((packed)) aacdec_out_ctx_st;
 
 typedef struct {
